revert unnecessary move
